requirements.txt
requests~=2.31.0
PyYAML~=6.0.1
pyproj~=3.6.0
tqdm~=4.65.0